Updated charts
The main committee page now has a chart of the full committee. Each of
the chair's direct reports now has a full chart of their branch also.
Lastly, charts from levels below the previously mentioned ones were
removed as they provided next to no value.

In addition to the changes above that can be seen on the site, a couple
of changes that are for those of us maintaining things are also
included:

- Helen's name was replaced with the new shortcode that will fill it in
  dynamically
- The mermaid bits have indentation in them to represent the actual
  structure to make them easier to read.